refactor: move is-visitor hook to reflect backend organization
